% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_optional_info.R
\name{add_optional_info}
\alias{add_optional_info}
\title{Add the next LeadsTo column}
\usage{
add_optional_info(benefits, add)
}
\arguments{
\item{benefits}{Benefits tibble}

\item{add}{vector of columns to add (including LeadsTo and Measure)}
}
\value{
benefits structure
}
\description{
Add a new LeadsTo column. A new benefits structure from create_benefits has
a column labelled LeadsTo1 which indicates which benefit/or impact it contributes
to in the next group. If a benefit leads to more than more benefit in the next
group this needs to be recorded in a new column.
}
\examples{
b <- create_benefits()
b <- add_optional_info(b, add = "LeadsTo")
}
