% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_benefits.R
\name{clean_benefits}
\alias{clean_benefits}
\title{Validate a benefits structure}
\usage{
clean_benefits(benefits)
}
\arguments{
\item{benefits}{Benefits tibble}
}
\value{
Benefits tibble
}
\description{
This function tidies a benefits map structure. Lines with blank Level or
Element are skipped, LeadsTo are converted back to integer
}
\examples{
b <- clean_benefits()
}
