% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_benefits.R
\name{create_benefits}
\alias{create_benefits}
\title{Create a skeleton benefits mapping structure}
\usage{
create_benefits(
  count = 1,
  levels = c("Activity", "Outcome", "Business change", "Benefit", "Objective", "Value")
)
}
\arguments{
\item{count}{Number of elements (benefits) for each level in the table. This can
be either an integer, in which case the same number number of elements is
created at each level, vector of the same length as levels to get different
numbers of elements at each level}

\item{levels}{Vector of Titles for level, or "PMI" as shorthand for
c("Output",
  "Outcome",
  "Intermediate benefit",
  "End benefit",
  "Impact")}
}
\value{
benefits with standard columns, as a tibble
}
\description{
This function makes it easier to get started with creating a benefits map by
creating a template data frame that can be edited or exported. Each row
include one element at one level (eg an output, outcome, benefit or impact) and
describes which level it is at and which element it leads to at the next level.
The function also creates a number of LeadsTo columns that shows for each element
which element in the next level is dependent on that particular element. This is
created with a complete combinatorial connection.
}
\examples{
b <- create_benefits()
b
b <- create_benefits(count = c(2, 4, 1),
                     levels = c("What we will do",
                               "What will happen",
                               "Why should anyone care"))
b
b <- create_benefits(count = c(1, 2, 1),
                     levels = c("Assumption",
                                "Initiative",
                                "Outcome"))
}
