% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_benefits.R
\name{map_benefits}
\alias{map_benefits}
\title{Create a DiagrammeR map of from a benefits structure}
\usage{
map_benefits(
  benefits,
  title = NA,
  splines = "line",
  width = 0.75,
  xscale = 1.5,
  yscale = 1,
  levels_textwrap = 12,
  elements_textwrap = 15,
  fillcolour_benefit = benefits_fill_colours,
  fontcolour_benefit = benefits_text_colours,
  fillcolour_disbenefit = disbenefits_fill_colour,
  fontcolour_disbenefit = disbenefits_text_colour,
  fillcolour_measure = measure_fill_colour,
  fontcolour_measure = measure_text_colour,
  linkcolour_benefit = "grey",
  linkcolour_disbenefit = "grey",
  fontname = "Calibri,Arial,Helvetica",
  fontname_title = "Calisto MT,Times New Roman",
  title_colour = "grey"
)
}
\arguments{
\item{benefits}{Benefits tibble}

\item{title}{Title for the map}

\item{splines}{How edges are represented (line, spline, ortho or curved)}

\item{width}{Proportion of available width to use for element boxes}

\item{xscale}{Scale factor for x-dimension}

\item{yscale}{Scale factor for y-dimension}

\item{levels_textwrap}{String length for wrapping text in headers}

\item{elements_textwrap}{String length for wrapping text in benefits}

\item{fillcolour_benefit}{one or more colours to fill benefit elements}

\item{fontcolour_benefit}{one or more colours for text}

\item{fillcolour_disbenefit}{one or more colours to fill disbenefit elements}

\item{fontcolour_disbenefit}{one or more colours for all disbenefit text elements}

\item{fillcolour_measure}{Colour to fill any measures defined}

\item{fontcolour_measure}{Font colour for any measures}

\item{linkcolour_benefit}{colour for links for benefit elements}

\item{linkcolour_disbenefit}{colour for links for all disbenefit elements}

\item{fontname}{fontname for nodes and edges}

\item{fontname_title}{fontname for title}

\item{title_colour}{Font colour}
}
\value{
graph Graph that can be used by DiagrammeR for render or export
}
\description{
Use the structure and just plot the boxes
}
\examples{
b <- create_benefits()
b
g <- map_benefits(b)
DiagrammeR::render_graph(g)
}
