% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_benefits.R
\name{validate_benefits}
\alias{validate_benefits}
\title{Validate a benefits structure}
\usage{
validate_benefits(benefits)
}
\arguments{
\item{benefits}{Benefits tibble}
}
\value{
logical
}
\description{
This function checks that a valid benefits tibble has been supplied
}
\examples{
b <- create_benefits()
if (validate_benefits(b)) print("Benefits Valid")
}
